Revert saving method migration
While shifting the saving methods out of the subclasses makes sense to
theoretically reduce code duplication, with the variation in the data
properties of the subclasses the code required to correctly generate and
parse the inputs to pass back to the superclass ends up essentially
duplicating the code anyway. The generic save methods have been removed
and responsibility placed back with the subclasses.

Revert commit 9127599793d98d326102dbaa1502940b68ed7d72