Add copy inheritence, lock getdate to UTC
Because we inherit from the handle superclass, creating copies of the
data object (b = a) will only create a new pointer to the same object,
not a copy of the object. MATLAB provides a generic object copying
superclass for handle classes.

Lock getdate to UTC so we don't even need to bother with timezones. They
suck.